Deliverable1/requirements.txt
psycopg2-binary==2.8.6
numpy==1.16.3